Move PasswordEncoder and TwigExtension to config folder
