+++ .env
@@ -5,6 +5,6 @@
     "MAILER_HOST": "smtp.example.org",
     "MAILER_PORT": 25,
     "MAILER_ENCRYPTION": "ssl",
-    "MAILER_USERNAME": "username",
+    "MAILER_USERNAME": "example@email.com",
     "MAILER_PASSWORD": "password"
 }
